改换使用map存储房间.

+++ src/main/java/com/artemis/football/model/BasePlayer.java
@@ -18,7 +18,7 @@ public class BasePlayer {
     protected Channel channel;
     protected String username;
     protected String nickname;
-    protected Long roomId;
+    protected Integer roomId;
     protected String teamName;
     protected Units units = new Units();
